Extract component to encapsulate presentation logic
This moves all of the view logic from `SelectPlacementSchoolForm` to a
new `SelectPlacementSchool::View` component.
